Move Triton layer norm test under HSTU
